@@ -15,22 +15,22 @@
     },
     "dependencies": {
         "@floating-ui/dom": "^1.7.1",
-        "@stellar/stellar-sdk": "^13.3.0",
+        "@stellar/stellar-sdk": "^14.2.0",
         "buffer": "^6.0.3",
-        "lucide-svelte": "^0.513.0",
+        "lucide-svelte": "^0.544.0",
         "passkey-kit": "^0.10.24",
         "svelte-persisted-store": "^0.12.0"
     },
     "devDependencies": {
         "@skeletonlabs/skeleton": "^2.11.0",
         "@skeletonlabs/tw-plugin": "^0.4.1",
-        "@sveltejs/adapter-auto": "^6.0.1",
-        "@sveltejs/kit": "^2.21.2",
-        "@sveltejs/vite-plugin-svelte": "^5.1.0",
+        "@sveltejs/adapter-auto": "^6.1.0",
+        "@sveltejs/kit": "^2.43.5",
+        "@sveltejs/vite-plugin-svelte": "^6.2.1",
         "@tailwindcss/forms": "^0.5.10",
         "@tailwindcss/typography": "^0.5.16",
         "@types/eslint": "^9.6.1",
-        "@types/node": "^22.15.29",
+        "@types/node": "^24.5.2",
         "@typescript-eslint/eslint-plugin": "^8.33.1",
         "@typescript-eslint/parser": "^8.33.1",
         "autoprefixer": "^10.4.21",
@@ -43,12 +43,12 @@
         "postcss": "^8.5.4",
         "prettier": "^3.5.3",
         "prettier-plugin-svelte": "^3.4.0",
-        "svelte": "^5.33.14",
-        "svelte-check": "^4.2.1",
+        "svelte": "^5.39.6",
+        "svelte-check": "^4.3.2",
         "tailwindcss": "^3.4.17",
         "tslib": "^2.8.1",
-        "typescript": "^5.8.3",
-        "vite": "^6.3.5",
+        "typescript": "^5.9.2",
+        "vite": "^7.1.7",
         "vite-plugin-tailwind-purgecss": "^0.3.5",
         "ye_olde_guestbook": "file:packages/ye_olde_guestbook"
     }
