@@ -26,7 +26,7 @@
     <div class="flex flex-col bg-surface-100-800-token space-y-4 p-3 md:p-4">
         <div class="grid grid-cols-[auto_1fr_auto] gap-2 md:gap-8">
             <div class="md:!hidden self-center">
-                <button class="btn-icon btn-icon-sm" on:click={() => drawerStore.open()}>
+                <button class="btn-icon btn-icon-sm" onclick={() => drawerStore.open()}>
                     <Menu />
                 </button>
             </div>
