@@ -1,5 +1,10 @@
 <script lang="ts">
-    import { page } from '$app/stores';
+    import { page } from '$app/state';
+    interface Props {
+        children?: import('svelte').Snippet;
+    }
+
+    let { children }: Props = $props();
 </script>
 
 <ol class="breadcrumb">
@@ -7,7 +12,7 @@
     <li class="crumb-separator" aria-hidden="true">/</li>
     <li class="crumb"><a class="anchor" href="/read">Read</a></li>
     <li class="crumb-separator" aria-hidden="true">/</li>
-    <li>Message {$page.params.id}</li>
+    <li>Message {page.params.id}</li>
 </ol>
 
-<slot />
+{@render children?.()}
