@@ -1,4 +1,6 @@
 <script lang="ts">
+    import { run } from 'svelte/legacy';
+
     import {
         seAccountLink,
         seContractLink,
@@ -6,24 +8,30 @@
         seTransactionLink,
     } from '$lib/stellarExpert';
     import TruncatedAddress from './TruncatedAddress.svelte';
-    export let target: string | number;
-    let targetHref: string;
-
-    $: switch (typeof target) {
-        case 'number':
-            targetHref = seLedgerLink(target);
-            break;
-        case 'string':
-            targetHref = target.startsWith('C')
-                ? seContractLink(target)
-                : target.startsWith('G')
-                  ? seAccountLink(target)
-                  : seTransactionLink(target);
-            break;
-        default:
-            targetHref = 'https://stellar.expert';
-            break;
+    interface Props {
+        target: string | number;
     }
+
+    let { target }: Props = $props();
+    let targetHref: string = $state();
+
+    run(() => {
+        switch (typeof target) {
+            case 'number':
+                targetHref = seLedgerLink(target);
+                break;
+            case 'string':
+                targetHref = target.startsWith('C')
+                    ? seContractLink(target)
+                    : target.startsWith('G')
+                      ? seAccountLink(target)
+                      : seTransactionLink(target);
+                break;
+            default:
+                targetHref = 'https://stellar.expert';
+                break;
+        }
+    });
 </script>
 
 <a href={targetHref} target="_blank">
