@@ -1,18 +1,26 @@
 <script lang="ts">
+    import { run } from 'svelte/legacy';
+
     import { SlideToggle } from '@skeletonlabs/skeleton';
     import GuestbookMessage from '$lib/components/GuestbookMessage.svelte';
 
     import type { PageData } from './$types';
-    export let data: PageData;
-
-    let sortNewestFirst = true;
-    let messages = data.messages;
-
-    $: if (sortNewestFirst) {
-        messages = messages.sort((a, b) => b.ledger - a.ledger);
-    } else {
-        messages = messages.sort((a, b) => a.ledger - b.ledger);
+    interface Props {
+        data: PageData;
     }
+
+    let { data }: Props = $props();
+
+    let sortNewestFirst = $state(true);
+    let messages = $state(data.messages);
+
+    run(() => {
+        if (sortNewestFirst) {
+            messages = messages.sort((a, b) => b.ledger - a.ledger);
+        } else {
+            messages = messages.sort((a, b) => a.ledger - b.ledger);
+        }
+    });
 </script>
 
 <div class="flex flex-col md:flex-row justify-start md:justify-between space-y-4">
